Add bfloat16 autocast and BF16 benchmarking sweep
Enable bfloat16 autocast across benchmarks and run FP32 and BF16 sweeps concurrently. Changes include: switch example autocast to torch.bfloat16 in autocast.py; add autocast_bfloat16 flag to benchmark() and wrap warmup/trial runs in a bfloat16 autocast context (using contextlib.nullcontext when disabled) in cs336_systems/benchmarking.py; extend benchmark.py with _spawn_sweep and _gather helpers, propagate autocast_bfloat16 to spawned runs, and update main to launch and display both FP32 and BF16 result tables; update embedded sample results. Also update modal GPU target from H100 to H200 in cs336_systems/modal.py.

+++ autocast.py
@@ -30,7 +30,7 @@ def autocast():
     model = ToyModel(10, 10).to("cuda")
     x = torch.randn(10, 10, dtype=torch.float32).to("cuda")
     target = torch.randint(0, 10, (10,), device="cuda")
-    with torch.autocast(device_type="cuda", dtype=torch.float16):
+    with torch.autocast(device_type="cuda", dtype=torch.bfloat16):
         logits = model(x)
         loss = torch.nn.functional.cross_entropy(logits, target)
         print("loss", loss.dtype)
